requirements/dev.txt
-r requirements.txt
pytest 
httpx
uvicorn
fastapi-debug-toolbar
